Flatten `src/errors` module and remove `experimental/bard` feature
Flattened `src/errors/` directory structure into a single `src/errors.rs` file to reduce complexity, as the module size (<300 lines) did not justify a subdirectory.
Removed `src/experimental/bard.rs` and the `bard` CLI command, as the "whimsical English narrative" feature was deemed non-essential bloat.
Removed all references to `bard` in `src/tools/cli.rs`, `src/tools/runner.rs`, and `src/main.rs`.
Deleted `tests/bard_coverage.rs`.

This adheres to the KISS and YAGNI principles.
